chore: rename role to team members
